test: moved a test into correct folder
