+++ src/tests/unit/test_tree_functionality/test_tree_node_to_ref_dict.py
@@ -10,12 +10,84 @@ from tests.unit.test_tree_functionality.mock_data_for_tree_tests.get_node_for_tr
 from tests.unit.test_tree_functionality.mock_data_for_tree_tests.mock_blueprint_provider_for_tree_tests import (
     BlueprintProvider,
 )
+from tests.unit.test_tree_functionality.mock_data_for_tree_tests.mock_document_service_for_tree_tests import (
+    mock_document_service,
+)
 from tests.unit.test_tree_functionality.mock_data_for_tree_tests.mock_storage_recipe_provider import (
     mock_storage_recipe_provider,
 )
 
 
 class TreeNodeToRefDictTestCase(unittest.TestCase):
+    def test_tree_node_to_ref_dict(self):
+        # Arrange
+        engine_package_content_bp_attribute = BlueprintAttribute(
+            name="content", attribute_type="object", type="dmss://system/SIMOS/BlueprintAttribute"
+        )
+        engine_entity_ref = {
+            "address": "$123",
+            "type": SIMOS.REFERENCE.value,
+            "referenceType": REFERENCE_TYPES.LINK.value,
+        }
+
+        engine_package_entity = {
+            "_id": "26d94353-3ff0-4b7d-bf06-86f006dd6f7b",
+            "type": "dmss://system/SIMOS/Package",
+            "name": "EnginePackage",
+            "isRoot": False,
+            "content": [engine_entity_ref],
+        }
+        engine_package_blueprint_attribute = BlueprintAttribute(
+            name="Package",
+            attribute_type="dmss://system/SIMOS/Package",
+            type="dmss://system/SIMOS/BlueprintAttribute",
+        )
+        engine_package_node = Node(
+            key="Package",
+            entity=engine_package_entity,
+            attribute=engine_package_blueprint_attribute,
+            blueprint_provider=mock_document_service.get_blueprint,
+            recipe_provider=None,
+        )
+
+        engine_blueprint_attribute = BlueprintAttribute(
+            name="content",
+            attribute_type=SIMOS.REFERENCE.value,
+            type="dmss://system/SIMOS/BlueprintAttribute",
+            contained=False,
+        )
+        engine_ref_node = Node(
+            key="0",
+            entity=engine_entity_ref,
+            attribute=engine_blueprint_attribute,
+            blueprint_provider=mock_document_service.get_blueprint,
+            recipe_provider=None,
+            uid=engine_entity_ref["address"],
+        )
+
+        content = ListNode(
+            key="content",
+            attribute=engine_package_content_bp_attribute,
+            entity=[engine_entity_ref],
+            blueprint_provider=None,
+            recipe_provider=None,
+        )
+        content.children = [engine_ref_node]
+        engine_package_node.children = [content]
+
+        content.parent = engine_package_node
+        engine_ref_node.parent = content.children[0]
+
+        # Act
+        engine_package_dict = tree_node_to_ref_dict(engine_package_node)
+
+        # Assert
+        assert engine_package_dict["content"][0] == {
+            "address": "$123",
+            "type": SIMOS.REFERENCE.value,
+            "referenceType": REFERENCE_TYPES.LINK.value,
+        }
+
     def test_tree_node_to_ref_dict_2(self):
         form_node = get_form_example_node()
         form_dict = tree_node_to_ref_dict(form_node)
